+++ include/pinocchio/bindings/python/multibody/geometry-data.hpp
@@ -13,7 +13,7 @@
 #include "pinocchio/bindings/python/utils/printable.hpp"
 #include "pinocchio/bindings/python/utils/copyable.hpp"
 #include "pinocchio/bindings/python/utils/std-vector.hpp"
-#include "pinocchio/bindings/python/utils/registration.hpp"
+
 #include "pinocchio/bindings/python/serialization/serializable.hpp"
 
 #if EIGENPY_VERSION_AT_MOST(2, 8, 1)