+++ include/pinocchio/bindings/python/multibody/geometry-object.hpp
@@ -12,7 +12,7 @@
 
 #include "pinocchio/bindings/python/utils/address.hpp"
 #include "pinocchio/bindings/python/utils/copyable.hpp"
-#include "pinocchio/bindings/python/utils/registration.hpp"
+
 #include "pinocchio/bindings/python/utils/deprecation.hpp"
 #include "pinocchio/bindings/python/utils/pickle.hpp"
 #include "pinocchio/bindings/python/serialization/serializable.hpp"